# src/main/resources/api/swagger.yaml
...
swagger: 2.0
info:
  description: Hello World API
  version: 1.0
  title: hello-world
  contact:
    name: Arquitectura Software
    url: none
    email: ignacio.lopezv@santalucia.es
  x-dominio: Gadgets
  x-desc-dominio:
    schema:
      $ref: '#/definitions/desc-dominio'
host: apis.santalucia.es
basePath: /hello-world/v1
schemes:
- https
consumes:
- application/json
produces:
- application/json
security:
- oauthApplication: []
paths:
...
  /agencias/list:
    get:
      tags:
      - list-agencias
      summary: recovery agencias list
      description: recovery agencias list
      operationId: getAgenciasList
      x-spring-paginated: true
      parameters:
      - name: X-Request-ID
        in: header
        description: Identificador único de la petición
        required: false
        type: string
        format: uuid
        x-example: AVDGSGRT-DHUEYR74R-SJDK75EJDSK
      responses:
        "200":
          description: ok response
          schema:
            type: array
            items:
              $ref: '#/definitions/agencia'
...
securityDefinitions:
  oauthApplication:
    type: oauth2
    tokenUrl: https://sso-sldev.santalucia.es/auth/realms/stl-int/protocol/openid-connect/token
    flow: application
definitions:
...
  agencia:
    type: object
    properties:
      dprov:
        type: string
        example: "Madrid"
        description: Provincia de la agencia
        maxLength: 255
      dpoblaci:
        type: string
        example: "Alicante"
        description: Poblacion de la agencia
        maxLength: 255
      dagencia:
        type: string
        example: "Nombre agencia"
        description: Nombre agencia
        maxLength: 255
      dtvia:
        type: string
        example: "calle"
        description: tipo de via
        maxLength: 255
      ddomici:
        type: string
        example: "lepanto 5"
        description: domicilio de la agencia
        maxLength: 255
...